refactor(transformations): consolidate TransformationType enum to single source of truth
- Created canonical TransformationType enum in app/models/transformation.py
- Updated app/schemas/transformation.py to import from models
- Updated app/services/transformation_engine/transformation_engine.py to import from models
- Added DATA_LOSS_THRESHOLD_PERCENT constant (50%) for consistency
- Updated test to use constant and clearer documentation
- TransformationStep validator now derives allowed_types from enum

This eliminates enum duplication that could cause runtime errors when
types diverge between schema and engine.
